requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.3.1
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==5.0.7
django-dirtyfields==1.9.3
django-silk==5.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
Faker==26.0.0
gprof2dot==2024.6.6
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
msgpack==1.0.8
packaging==24.1
pluggy==1.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
redis==5.0.7
referencing==0.35.1
rpds-py==0.19.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
tqdm==4.66.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
zope.interface==6.4.post2
